+++ lib/ui/energies/EnergyTile.dart
